Fix motherLi logic. Prevent whitespace nodes, notice, multiple current indicators.
Fix motherLi logic: The ul containing the submenu lis (children) should
go within the motherLi. It was rendered as separate li that showed as
dot or space inbetween menu entries as only the ul had display: none.

Prevent whitespace nodes: return '</li>."\n" resulted in unwanted
whitespace between lis of the menu. At least recent versions of firefox
developer tools show these nodes.

Prevent notice on __ as explode on _ resulted in an empty word that was
not handled correctly in toFairy function.

Prevent multiple current indicators: With recursion the folders' entries
also show up in the menu and are classified as `current` because they
contain an index.htm or home.htm et alia which by default are indicating
the start page ($startDE, $startEN).